Fix display of number with scientific notation
